requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
langchain>=0.3.0
langchain-google-genai>=2.1.0
langchain-experimental>=0.3.0
pandas>=2.2.0
matplotlib>=3.9.0
seaborn>=0.13.0
streamlit>=1.40.0
requests>=2.31.0
tabulate>=0.9.0
python-dotenv>=1.0.0
